pi-firmware/requirements.txt
pyserial>=3.4
PyYAML>=5.3.1
luma.lcd>=2.9.0
pillow>=9.0.0
